refactor(struct): fix review issues
