+++ libraries/struct/src/field/index.ts
@@ -1,0 +1,3 @@
+export * from "./factory.js";
+export * from "./serialize.js";
+export * from "./types.js";
