+++ libraries/struct/src/index.ts
@@ -14,7 +14,7 @@ export * from "./bipedal.js";
 export * from "./buffer.js";
 export * from "./concat.js";
 export * from "./extend.js";
-export * from "./field.js";
+export * from "./field/index.js";
 export * from "./number.js";
 export * from "./readable.js";
 export * from "./string.js";
